Drop support for LTS-14 and LTS-15
We are no longer going to support the stack-lts-14 and stack-lts-15
resolvers. The new minimum is LTS-16.33, which targets the GHC v8.8.4.

Most packages only see a patch update, a few have some feature updates.
generic-lens is the biggest change, jumping a full major version.

However, this will allow us to drop support for lens in the mainline
SDK, as requested [in a previous pr][1].

[1]: https://github.com/launchdarkly/haskell-server-sdk/pull/34